Complte course ui and bug fix

+++ src/lk/ijse/studentmanagementsystem/business/custom/CourseBO.java
@@ -11,9 +11,9 @@ public interface CourseBO extends SuperBO {
 
     public List<CourseTM> getAllCourse() throws Exception;
     public Course findSubject (String id) throws Exception;
-    public  boolean saveCourse(String id, String name, BigDecimal fee, String description)throws Exception;
+    public  boolean saveCourse(String id, String name, double fee, String description)throws Exception;
     public  boolean deleteCourse(String courseId)throws Exception;
-    public  boolean updateCourse(String name, BigDecimal fee, String description,String id)throws Exception;
+    public  boolean updateCourse(String name, double fee, String description,String id)throws Exception;
     public  String getNewCourseId()throws Exception;
